+++ lib/protocols.py
@@ -2,4 +2,4 @@ from typing import Protocol
 
 
 class Movable(Protocol):
-    distanceTravelled: int
+    distanceTravelled: float
